feat:Transfer logic to login and signup
